requirement.txt
certifi==2019.6.16
chardet==3.0.4
DingtalkChatbot==1.3.0
Django==1.11.8
dnspython==1.16.0
idna==2.8
IPy==1.0
mysqlclient==1.4.2.post1
psutil==5.6.2
PyMySQL==0.9.3
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
